core/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
django-cors-headers==3.5.0
drf-social-oauth2==1.0.8
Pillow==9.2.0